llvm:  Fix inconsistent uses of .cfi_sections
Error: inconsistent uses of .cfi_sections

From the 2.27 binutils assemblers. It seems that 2.27 doesn't like the

.cfi_sections .debug_frame

directive following previous .cfi directives. The assemblers seem to be happy if the .cfi_sections directive precedes any other .cfi directive. Is this a bug in binutils or LLVM? It looks as if a fix might be to move the generation of the .cfi_sections directive from endModule() to somewhere else (maybe a non-existant beginModule()?).

As a work-around I modified DwarfCFIException.cpp and ARMException.cpp to send the .cfi_sections directive before the first function is processed. Seems to work.

-Rich

Signed-off-by: Khem Raj <raj.khem@gmail.com>